+++ nodeserver/constants.js
@@ -1,17 +1,6 @@
 // constants.js
 const envSwitcher = require('./envSwitcher');
 
-const baseUrl = envSwitcher.baseUrl;
-const urlMap = {
-    'bewaesserung/switch/stefanNord': `${baseUrl}/set/tuya.0.51050522600194fed14c.1`,
-    'bewaesserung/switch/stefanOst': `${baseUrl}/set/tuya.0.51050522600194fed14c.2`,
-    'bewaesserung/switch/lukasSued': `${baseUrl}/set/tuya.0.51050522600194fed14c.3`,
-    'bewaesserung/switch/lukasWest': `${baseUrl}/set/tuya.0.51050522600194fed14c.4`,
-    'bewaesserung/switch/alle': `${baseUrl}/set/tuya.0.51050522600194fed14c.13`,
-    'markise/switch/haupt' : `${baseUrl}/set/tuya.0.8407060570039f7fa6d2.1`,
-    // add other mappings
-}
-
 // Define a mapping of topics to task enabler keys
 const topicToTaskEnablerKey = {
     'stefanNord': 'Stefan_Nord',
@@ -23,4 +12,4 @@ const topicToTaskEnablerKey = {
 
 const mqttBrokerUrl = envSwitcher.mqttBrokerUrl;
 
-module.exports = { mqttBrokerUrl, urlMap, topicToTaskEnablerKey };
+module.exports = { mqttBrokerUrl, topicToTaskEnablerKey };